Add pavicrawler wps process & major improvements
The pavicrawler process will allow adding NetCDF metadata to the Solr
database.
The pavicsearch process was updated with a working Solr query
construction from the inputs.
The pavicsupdate process was updated to take into account the
id composed of the source and url fields and other particularities
of the birdhouse-solr implementation.