Change update method to behave as if adding a full document
Partial updates do not seem to work with birdhouse-solr and using the
source and url fields to search for documents to update also fails.
Therefore the id field is now used to search for the document, then
all its current information is kept, adding the requested updates,
and sending an add request with all the information that will overwrite
the previous entry, but keep the same id.