Fix use of datetime vs netcdftime
The time_vectors_to_datetimes function was not compatible with
typical NetCDF4 usage as a function of calendar. When the calendar is
gregorian, datetime.datetime objects should be returned, while for the
other calendars, netCDF4.netcdftime.datetime should be used.
Also added support for the microsecond concept present in both
datetime and netcdftime.